requirements.txt
telethon==1.36.0
cryptography==41.0.7
python-dotenv==1.0.0
eth-account==0.10.0
nest-asyncio==1.5.8
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
hyperliquid-python-sdk>=0.15.0
aiohttp
pytest-asyncio
sentry-sdk
web3
